requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
gitdb==4.0.11
GitPython==3.1.41
pillow==11.0.0
setuptools==69.0.3
shortuuid==1.0.13
smmap==5.0.1
sqlparse==0.5.1
